Change (int) refresh_rate to SDL_RefreshRate with numerator and denominator

+++ include/SDL3/SDL_test_common.h
@@ -78,7 +78,7 @@ typedef struct
     int logical_h;
     float scale;
     int depth;
-    int refresh_rate;
+    double refresh_rate;
     int num_windows;
     SDL_Window **windows;
